backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.54.5
pydantic==2.10.1
python-dotenv==1.0.1
